requirements.txt
python-telegram-bot==20.3
flask
stripe
python-dotenv
gunicorn
